refactor: remove unused components and implement ErrorPage for better error handling
