--- src/components/RouteGuard.jsx
@@ -1,19 +1,0 @@
-import { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
-
-const RouteGuard = ({ children }) => {
-  const { isAuthenticated, loading } = useAuth();
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    // Only redirect after initial loading is complete
-    if (!loading && !isAuthenticated) {
-      navigate("/login");
-    }
-  }, [isAuthenticated, loading, navigate]);
-
-  return children;
-};
-
-export default RouteGuard;
